Moving main class to swing gui
